# Improve MiqQueue#wait_times_by_role