`MiqQueue#wait_times_by_role` determines the wait times in the queue for each role.

It currently brings back every row in `miq_queue` and ruby filters out all but 2 records per role.

This PR does that filtering in the database, so only a record per role is transferred from the db to ruby.

While collecting metrics, I realized there was an alternative that would perform much better. In my tests, all queries returned the same data, but the alternative does have the potential to return slightly different results.

Here is a summary of the solutions. I included the time spent in the database for 1 and 21k rows. I also displayed the time spent in ruby for the 21k rows case. Note that the db time increases in this PR.

Method|ruby time 21k|db 1 time| db 21k time |rows returned (21k)|SQL Complexity|ruby complexity
------|-------------|---------|------------|-------|--------------|---
Before|  0.76s      | 0.302   |  73.649ms  |21,000 |good          |good
After |  **0.04s**      | 0.463   | 945.195ms  |**5**      |bad           |mid
Alt   |  **0.02s**      | 0.180   |  20.898ms  |**5**      |good          |good

TL;DR: My suggestion is to go with the Alt code, but if we need the identical results, the new code looks better.

/cc @chessbyte @Fryguy @jrafanie @gmcculloug Found this old code from when I was working on `MiqQueue`. Not sure the frequency of this method but it had stood out because it sorts all of `miq_queue` and it then brings back every row. Since I had already written the code, thought I'd at least put in a PR.


**Only read on if you have a strong stomach**

---
---
Algorithm overviews

TL;DR: My goal was to reduce the amount of data over the wire, at the cost of adding processing to the database. The alt view happens to reduce both.

Do note that the algorithms work because sorting by `id` is essentially the same thing as sorting by `created_on`.


Before:

1. sort miq_queue by priority and id
2. db -> ruby one record for every row
3. group by role
4. for each role: filter for first (least amount of time) and last (most amount of time)

After:

1. sort miq_queue by priority and id
2. group by role
3. for each role: filter for first (least amount of time) and last (most amount of time)
4. db -> ruby one record for every role

Alt:

1. don't sort
2. group by role
3. bring back min / max created_on (not taking priority into account)
4. db -> ruby one record for every role

---
The ruby code

TL;DR: alt/before are the best, after is worst.

Included the alt below. The before and after is in the PR.

After:
```ruby
w="(partition by role order by priority, id range between unbounded preceding and unbounded following)"
now = Time.now.utc
dates = where(:state => STATE_READY)
  .select("distinct role,
           first_value(created_on) over #{w} as next_in_queue,
           last_value(created_on) over #{w} as last_in_queue")
dates.each.with_object({}) do |o, h|
  h[o.role] = {
    :next => (now - o['last_in_queue'].to_datetime),
    :last => (now - o['next_in_queue'].to_datetime)
  }
end
```

Alt:
```ruby
MiqQueue.where(:state => %w(state ready)).group(:role).pluck("min(created_on)","max(created_on)")
```

---
What do the queries look like?

TL;DR: alt/beforeare the best, after is worst.

Before:
```
SELECT created_on, role
from miq_queue
where miq_queue.state in ('state', 'ready')
order by priority, id;
```

After:
```
SELECT DISTINCT role,
first_value(created_on) OVER w AS next_in_queue,
last_value(created_on) OVER w AS last_in_queue
FROM "miq_queue"
WHERE "miq_queue"."state" in ('state', 'ready')
WINDOW w AS (PARTITION BY role ORDER BY priority, id RANGE BETWEEN unbounded preceding and unbounded following);
```

Alt:
```
SELECT role, min(created_on), max(created_on)
FROM "miq_queue"
WHERE miq_queue.state in ('state', 'ready')
group by role;
```

---
Processing for 1 row in the database (1 role)

TL;DR: for a single row, alt is the best, but others are close.

Before:
```
 Sort  (cost=29.44..29.48 rows=14 width=536) (actual time=0.272..0.272 rows=1 loops=1)
   Sort Key: priority, id
   Sort Method: quicksort  Memory: 25kB
   ->  Seq Scan on miq_queue  (cost=0.00..29.18 rows=14 width=536) (actual time=0.045..0.258 rows=1 loops=1)
         Filter: ((state)::text = ANY ('{state,ready}'::text[]))
 Total runtime: 0.302 ms
```

After:
```
 HashAggregate  (cost=29.90..30.04 rows=14 width=536) (actual time=0.163..0.163 rows=1 loops=1)
   ->  WindowAgg  (cost=29.44..29.79 rows=14 width=536) (actual time=0.156..0.156 rows=1 loops=1)
         ->  Sort  (cost=29.44..29.48 rows=14 width=536) (actual time=0.132..0.132 rows=1 loops=1)
               Sort Key: role, priority, id
               Sort Method: quicksort  Memory: 25kB
               ->  Seq Scan on miq_queue  (cost=0.00..29.18 rows=14 width=536) (actual time=0.010..0.109 rows=1 loops=1)
                     Filter: ((state)::text = ANY ('{state,ready}'::text[]))
 Total runtime: 0.463 ms
```
Alt:
```
 HashAggregate  (cost=29.28..29.42 rows=14 width=524) (actual time=0.123..0.123 rows=1 loops=1)
   ->  Seq Scan on miq_queue  (cost=0.00..29.18 rows=14 width=524) (actual time=0.011..0.118 rows=1 loops=1)
         Filter: ((state)::text = ANY ('{state,ready}'::text[]))
 Total runtime: 0.180 ms
```

---
Processing for 21,170 rows in the database (5 roles)

TL;DR: In the database, alt is 3x faster than before, with after being 10x slower. But alt and after bring back 5 rows while before brings back 21k rows.

Before:
```
 Sort  (cost=3136.58..3189.51 rows=21171 width=41) (actual time=61.713..68.726 rows=21169 loops=1)
   Sort Key: priority, id
   Sort Method: external merge  Disk: 1192kB
   ->  Seq Scan on miq_queue  (cost=0.00..960.96 rows=21171 width=41) (actual time=0.010..14.997 rows=21169 loops=1)
         Filter: ((state)::text = ANY ('{state,ready}'::text[]))
         Rows Removed by Filter: 26
 Total runtime: 73.649 ms
```
After:
```
 Unique  (cost=5841.47..6053.18 rows=21171 width=41) (actual time=932.197..944.082 rows=5 loops=1)
   ->  Sort  (cost=5841.47..5894.40 rows=21171 width=41) (actual time=932.196..937.559 rows=21169 loops=1)
         Sort Key: role, (first_value(created_on) OVER (?)), (last_value(created_on) OVER (?))
         Sort Method: external sort  Disk: 1368kB
         ->  WindowAgg  (cost=3136.58..3665.85 rows=21171 width=41) (actual time=591.635..739.942 rows=21169 loops=1)
               ->  Sort  (cost=3136.58..3189.51 rows=21171 width=41) (actual time=582.015..615.822 rows=21169 loops=1)
                     Sort Key: role, priority, id
                     Sort Method: external merge  Disk: 1192kB
                     ->  Seq Scan on miq_queue  (cost=0.00..960.96 rows=21171 width=41) (actual time=0.018..14.706 rows=21169 loops=1)
                           Filter: ((state)::text = ANY ('{state,ready}'::text[]))
                           Rows Removed by Filter: 26
 Total runtime: 945.195 ms
```

Alt:
```
 HashAggregate  (cost=1119.75..1119.79 rows=4 width=29) (actual time=20.863..20.865 rows=5 loops=1)
   ->  Seq Scan on miq_queue  (cost=0.00..960.96 rows=21171 width=29) (actual time=0.008..12.088 rows=21169 loops=1)
         Filter: ((state)::text = ANY ('{state,ready}'::text[]))
         Rows Removed by Filter: 26
 Total runtime: 20.898 ms
```