@Fryguy 2015-06-19
- It's literally only called [here](https://github.com/ManageIQ/manageiq/blob/0699eb4345178cf6c4118faa81df275f0fe1b487/vmdb/app/models/miq_queue_analytic.rb#L40) and also in the spec for the method itself.
- I literally read the whole thing, and I'm a little confused by the actions taken here with respect to the evidence.  There is a lot of stuff about the "after" version, but all of the numbers points to the "alt" version being the best.  It's faster, easier to read both in Ruby and SQL, is cross-database, and returns less rows.  Yet, this PR has the code for the "after" instead.  Why?  Just make the PR be the "alt" and ignore everything about the "after".  It's interesting, and perhaps a good learning experience for the WINDOW functions, but basically is just noise in this PR.