@Fryguy 2015-06-19
Also, given that caller, who adds it's own conditions, would the WINDOWing functions still work?  I think so, but again, that's the "after" code, and the "alt" is better anyway.