@kbrock 2015-06-20
@Fryguy thanks. I had made this branch a year and a half ago. I only added the alt when I saw the db cost of the after query.

I was concerned that a low priority task could skew the times, but that is probably theoretical only. I'll change.

Considering that it probably isn't called that often, I already spent too much time just in writing it up.