@jrafanie 2015-06-22
@kbrock Not really, I forgot about that method and didn't know if we used that method often enough to optimize it.